Always display AA flood geojson button
